Absorb test, envTest into jestTest
