+++ package.json
@@ -30,9 +30,9 @@
     "test": "npm run test-backend && npm run test-frontend",
     "test-backend": "         dotnet test Backend.Tests/Backend.Tests.csproj",
     "test-backend:coverage": "dotnet test Backend.Tests/Backend.Tests.csproj /p:CollectCoverage=true /p:CoverletOutputFormat=lcov /p:Threshold=77",
-    "test-frontend": "         node               scripts/test.js",
-    "test-frontend:coverage": "node               scripts/test.js --coverage --watchAll=false",
-    "test-frontend:debug": "   node --inspect-brk scripts/test.js --runInBand --no-cache",
+    "test-frontend": "         node               scripts/jestTest.js",
+    "test-frontend:coverage": "node               scripts/jestTest.js --coverage --watchAll=false",
+    "test-frontend:debug": "   node --inspect-brk scripts/jestTest.js --runInBand --no-cache",
     "wordlist": "hunspell-reader words"
   },
   "dependencies": {
